Decoupled express from index so it can be exported for unit tests
